64-bit times
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@718 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ client/BufferedSocket.cpp
@@ -136,7 +136,7 @@ void BufferedSocket::threadConnect(const string& aAddr, uint16_t aPort, bool pro
 		return;
 	fire(BufferedSocketListener::Connecting());
 
-	uint32_t startTime = GET_TICK();
+	uint64_t startTime = GET_TICK();
 	if(proxy) {
 		sock->socksConnect(aAddr, aPort, CONNECT_TIMEOUT);
 	} else {
